add DocumentDAO, DocumentDAOImpl, DocumentService, DocumentController

+++ src/main/java/io/khasang/archivarius/dao/impl/DocumentDAOImpl.java
@@ -9,7 +9,6 @@ import org.hibernate.criterion.Restrictions;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Repository;
 
-import java.util.Date;
 import java.util.List;
 
 @Repository
@@ -23,58 +22,20 @@ public class DocumentDAOImpl implements DocumentDAO {
     }
 
     @Override
-    public void updateDocumentTitle(Document document) {
-        final String query = "SELECT title from Document WHERE id=:id";
-        String newTitle = (String) sessionFactory.
-                getCurrentSession().
-                createSQLQuery(query).
-                setParameter("id", document.getId()).
-                uniqueResult();
-        document.setTitle(newTitle);
-        sessionFactory.getCurrentSession().update(document);
-    }
-
-    @Override
-    public void updateDocumentStatus(Document document) {
-        final String query = "SELECT status from Document WHERE id=:id";
-        String newStatus = (String) sessionFactory.
-                getCurrentSession().
-                createSQLQuery(query).
-                setParameter("id", document.getId()).
-                uniqueResult();
-        document.setStatus(newStatus);
-        sessionFactory.getCurrentSession().update(document);
-    }
-
-    @Override
-    public void updateDocumentDeadline(Document document) {
-        final String query = "SELECT deadline from Document WHERE id=:id";
-        Date newDeadline = (Date) sessionFactory.
-                getCurrentSession().
-                createSQLQuery(query).
-                setParameter("id", document.getId()).
-                uniqueResult();
-        document.setDeadline(newDeadline);
-        sessionFactory.getCurrentSession().update(document);
-    }
-
-    @Override
-    public void updateDocumentDestination(Document document) {
-        final String query = "SELECT destination from Document WHERE id=:id";
-        String newDestination = (String) sessionFactory.
-                getCurrentSession().
-                createSQLQuery(query).
-                setParameter("id", document.getId()).
-                uniqueResult();
-        document.setDestination(newDestination);
-        sessionFactory.getCurrentSession().update(document);
-    }
-
-    @Override
-    public void deleteDocument(Document document) {
+    public void updateDocument(Document document) {
         final Session session = sessionFactory.getCurrentSession();
-        session.delete(document);
+        session.saveOrUpdate(document);
         session.flush();
+    }
+
+    @Override
+    public void deleteDocument(int id) {
+        final Session session = sessionFactory.getCurrentSession();
+        Document document = session.load(Document.class, new Integer(id));
+        if (document != null) {
+            session.delete(document);
+            session.flush();
+        }
     }
 
     @Override
@@ -102,7 +63,6 @@ public class DocumentDAOImpl implements DocumentDAO {
                 createCriteria(Document.class);
         criteria.add(Restrictions.eq("destination", destination));
         return (Document) criteria.uniqueResult();
-
     }
 
     @Override
